requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.3
Pillow==10.4.0
reportlab==4.0.7
openpyxl==3.1.2
